Created the sweep_from function
This function assumes that the input tile which it receives does not have a flag, nor a question mark. It receives the tile inputted by the user, if it has a mine, it makes the loss variable true.
If the input tile has a number other than 0, it simply opens up that tile. 
If the input tile has 0, it flood fills all the zeros and opens up other tiles accordingly. If any tile which was flagged or had question mark, but got opened during flood fill, then the function changes the status of is_flagged and is_q_marked arrays. The function also updates the value of num_swept.