Fixed some bugs
The clear function was clearing even the swept tiles like numbers with 0 and 1,2,... so on which should not be the case, Hence I wrote updated state inside if condition
Added the get_num_swept function in Minefield class and added printing statements which shows how many tiles are swept and how many flags are there.
Next, removed the is_flagged array and is_q_marked array and is_swept array from the two functions sweep_util() and sweep_from()
There was a bug in sweep_from function which swept the flagged tiles without changing the number of flags. Hence that is tackled. i.e. the sweep_from function now tackles falgged and q_marked tiles too if they are swept.
There was another bug in sweep_from function that if we flagged a non-zero neighbour and later swept it's zero neighbour, the non-zero neighbour tiled showed  swept_char i.e. 'o' which shuld not be the case. Hence I fixed it line 333-336.
Try running the code after I removed superfluous arrays from the sweep functions.